feat: multiple platform cpes (#203)
Previously, if the platform CPE was more complicated than a single "running
on/with" entry, grype-db would leave it blank. Now, when several "running 
on/with" entries are present, emit a row for each of them, in order to reduce false
positives caused by more permissive platform constraint.

Note that "running on/with" could be an application or platform CPE.
For example, Redis or OpenShift might be coded as a type "a" CPE (for
"application"), but might be a platform (displayed in the "running on or
with" section of the NVD UI). For these, consider them platforms and
emit a platform CPE.

Signed-off-by: Will Murphy <will.murphy@anchore.com>