Fail cache status on empty results (#257)
Previously, grype-db cache status would exit zero in a few situations that
should have failed validation. Specifically, if the cached results had zero rows,
or if "-p some-provider" was passed, but "some-provider" no cache for
"some-provider" was present on disk, the command would exit zero. Instead,
exit non-zero if a provider is explicitly requested on the command line but
missing, and add a --min-rows option that defaults to zero, so that providers
with zero rows fail validation. If zero rows is acceptable, "--min-rows -1" can
be passed.

Signed-off-by: Will Murphy <will.murphy@anchore.com>