fix: ensure normalization of all severity values (#259)
Ensure that all severity values get normalized.  Previously
normalization would be skipped for any records that weren't unknown,
were from the NVD namespace, or were not based on CVE ids.  In those
cases what should be skipped is defaulting to the NVD provider values
when unknown, but not normalizing to a set enumeration of values

Signed-off-by: Weston Steimel <commits@weston.slmail.me>