fix: emit fuzzy ranges rather than invalid ranges
Previously, if upstream data contained invalid version ranges for the
ecosystem, e.g. pip packages that do not have a PEP440 compliant
version, grype db would just emit a version constraint that couldn't be
parsed at runtime. Instead, emit a verion constraint of type "unknown"
to signal to grype that the fuzzy version matcher should be used in this
case.

Signed-off-by: Will Murphy <willmurphyscode@users.noreply.github.com>