chore: use latest released grype-db to publish DB
Otherwise, there's a concern that unreleased changes in the grype-db
repo will reference models in unreleased changes in grype. Basically,
this change makes it so that merging into grype-db won't affect the
published database until grype-db is released. This behavior is probably
what everyone expected anyway, and is how vunnel has always behaved.

Signed-off-by: Will Murphy <willmurphyscode@users.noreply.github.com>