ci: increase runner size and ubuntu concurrency
The ubuntu provider is consistently timing out. It should be rewritten
to use a more performant approach, but first today's Grype DB needs to
be published. Temporarily raise the amount of compute available to the
daily data sync jobs and increase the amount concurrency requested of
the vunnel ubuntu provider.

Signed-off-by: Will Murphy <willmurphyscode@users.noreply.github.com>